Squashed commit of the following:
commit df3078f76f9de5bed125fe71abdecd0f24f5d3d3
Author: Gabe Stocco <98900+gfs@users.noreply.github.com>
Date:   Mon Apr 13 20:14:43 2020 -0700

    Update Documentation (#379)

    * Update Docs

    * Add additional comments to be picked up by docfx

    * Fix docfx to see new lib location.

    * Clean up

    * Update index.md

    * Update docfx.json

    * Update CertificateObject.cs

    * Update CommandOptions.cs

    * Update CollectObject.cs

    * EventLogObjects now use a proper timestamp

    * Better document ComObject

    * use dllcharacteristics natively instead of tostring

    * Document EventLogObject

    * Collect AddressFamily and Transport Enums instead of strings.

    * More comments

    * Separate some classes into their own files.

    * Update Settings.cs

    * Modify how RowKeys work so objects could implement unique row keys per object if needed.

    * Performance improvements by shortcutting operators when appropriate.

    * Document + Clarity refactor of Analyzer.cs

    * Parsing bug in analyzer.cs

    * Update Analyzer.cs

    * Missing ;s

    * Build fixes.

    * Run GenerateDocfx

    * Fixes

    * Update README.md

    * Cache clause evaluations

    * Fix off by 1

    * Use our own Enum

    * Update CollectObject.cs

    * Update AnalyzerExpressionsTests.cs

    * Add Address Family back

    * Switch to newtonsoft.json

    * Update docs

    * Fix Serialization

    * Fix LGTM Issues

    * More LGTM Fixes

    * Use an actual random password for tests

    The pipeline systems require a certian password complexity that just a GUID wasn't matching.

    * Improve RunExternalCommand

    * Reduce maximum length to avoid interactive prompt

    * Update CollectorTests.cs

    * Update pr-validation.yml

commit 7eea9d741366fb1baa1b227fac10bf74fac060ab
Author: Gabe Stocco <98900+gfs@users.noreply.github.com>
Date:   Mon Apr 13 18:00:13 2020 -0700

    Start Releasing 2.2 (#378)

    * Update release.yml

    * Update README.md

    * Build symbols nupkg

    * Publish symbols with nupkgs

    * Update Dependencies

    * Update GlobalSuppressions.cs

commit 4b27757ab200629a399a43f652405894e258609a
Merge: 3d8231d 4a14c43
Author: Gabe Stocco <98900+gfs@users.noreply.github.com>
Date:   Fri Apr 10 22:00:12 2020 -0700

    Merge branch 'release/v2.2'

commit 3d8231d971c3baac7b8894659923fe0caf28fa56
Author: Gabe Stocco <98900+gfs@users.noreply.github.com>
Date:   Fri Apr 10 22:00:12 2020 -0700

    Set version to '2.3-alpha'

+++ Lib/Collectors/CertificateCollector.cs
@@ -45,8 +45,10 @@ namespace AttackSurfaceAnalyzer.Collectors
                             var obj = new CertificateObject(
                                 StoreLocation: storeLocation.ToString(),
                                 StoreName: storeName.ToString(),
-                                Certificate: new SerializableCertificate(certificate),
-                                Pkcs7: certificate.Export(X509ContentType.Cert).ToString());
+                                Certificate: new SerializableCertificate(certificate))
+                            {
+                                Pkcs7 = Convert.ToBase64String(certificate.Export(X509ContentType.Cert))
+                            };
                             DatabaseManager.Write(obj, RunId);
                         }
 
@@ -68,28 +70,31 @@ namespace AttackSurfaceAnalyzer.Collectors
         {
             try
             {
-                var result = ExternalCommandRunner.RunExternalCommand("ls", new string[] { "/etc/ssl/certs", "-A" });
-
-                foreach (var _line in result.Split('\n'))
+                if (ExternalCommandRunner.RunExternalCommand("ls", "/etc/ssl/certs -A", out string result, out string _) == 0)
                 {
-                    Log.Debug("{0}", _line);
-                    try
+                    foreach (var _line in result.Split('\n'))
                     {
-                        using X509Certificate2 certificate = new X509Certificate2("/etc/ssl/certs/" + _line);
+                        Log.Debug("{0}", _line);
+                        try
+                        {
+                            using X509Certificate2 certificate = new X509Certificate2("/etc/ssl/certs/" + _line);
 
-                        var obj = new CertificateObject(
-                            StoreLocation: StoreLocation.LocalMachine.ToString(),
-                            StoreName: StoreName.Root.ToString(),
-                            Certificate: new SerializableCertificate(certificate),
-                            Pkcs7: certificate.Export(X509ContentType.Cert).ToString());
+                            var obj = new CertificateObject(
+                                StoreLocation: StoreLocation.LocalMachine.ToString(),
+                                StoreName: StoreName.Root.ToString(),
+                                Certificate: new SerializableCertificate(certificate))
+                            {
+                                Pkcs7 = Convert.ToBase64String(certificate.Export(X509ContentType.Cert))
+                            };
 
-                        DatabaseManager.Write(obj, RunId);
-                    }
-                    catch (Exception e)
-                    {
-                        Log.Debug("{0} {1} Issue creating certificate based on /etc/ssl/certs/{2}", e.GetType().ToString(), e.Message, _line);
-                        Log.Debug("{0}", e.StackTrace);
+                            DatabaseManager.Write(obj, RunId);
+                        }
+                        catch (Exception e)
+                        {
+                            Log.Debug("{0} {1} Issue creating certificate based on /etc/ssl/certs/{2}", e.GetType().ToString(), e.Message, _line);
+                            Log.Debug("{0}", e.StackTrace);
 
+                        }
                     }
                 }
             }
@@ -109,33 +114,42 @@ namespace AttackSurfaceAnalyzer.Collectors
         {
             try
             {
-                var result = ExternalCommandRunner.RunExternalCommand("security", new string[] { "find-certificate", "-ap", "/System/Library/Keychains/SystemRootCertificates.keychain" });
-                string tmpPath = Path.Combine(Directory.GetCurrentDirectory(), "tmpcert.pem");
-                string pkPath = Path.Combine(Directory.GetCurrentDirectory(), "tmpcert.pk12");
-
-                File.WriteAllText(tmpPath, result);
-                _ = ExternalCommandRunner.RunExternalCommand("openssl", new string[] { "pkcs12", "-export", "-nokeys", "-out", pkPath, "-passout pass:pass", "-in", tmpPath });
-
-                X509Certificate2Collection xcert = new X509Certificate2Collection();
-                xcert.Import(pkPath, "pass", X509KeyStorageFlags.DefaultKeySet);
-
-                File.Delete(tmpPath);
-                File.Delete(pkPath);
-
-                var X509Certificate2Enumerator = xcert.GetEnumerator();
-
-                while (X509Certificate2Enumerator.MoveNext())
+                if (ExternalCommandRunner.RunExternalCommand("security", "find-certificate -ap /System/Library/Keychains/SystemRootCertificates.keychain", out string result, out string _) == 0)
                 {
-                    var certificate = X509Certificate2Enumerator.Current;
+                    string tmpPath = Path.Combine(Directory.GetCurrentDirectory(), "tmpcert.pem");
+                    string pkPath = Path.Combine(Directory.GetCurrentDirectory(), "tmpcert.pk12");
 
-                    var obj = new CertificateObject(
-                        StoreLocation: StoreLocation.LocalMachine.ToString(),
-                        StoreName: StoreName.Root.ToString(),
-                        Certificate: new SerializableCertificate(certificate),
-                        Pkcs7: certificate.Export(X509ContentType.Cert).ToString());
+                    File.WriteAllText(tmpPath, result);
+                    if (ExternalCommandRunner.RunExternalCommand("openssl", $"pkcs12 -export -nokeys -out {pkPath} -passout pass:pass -in {tmpPath}",out string _, out string _) == 0)
+                    {
+                        X509Certificate2Collection xcert = new X509Certificate2Collection();
+                        xcert.Import(pkPath, "pass", X509KeyStorageFlags.DefaultKeySet); //lgtm [cs/hardcoded-credentials]
 
-                    DatabaseManager.Write(obj, RunId);
+                        File.Delete(tmpPath);
+                        File.Delete(pkPath);
+
+                        var X509Certificate2Enumerator = xcert.GetEnumerator();
+
+                        while (X509Certificate2Enumerator.MoveNext())
+                        {
+                            var certificate = X509Certificate2Enumerator.Current;
+
+                            var obj = new CertificateObject(
+                                StoreLocation: StoreLocation.LocalMachine.ToString(),
+                                StoreName: StoreName.Root.ToString(),
+                                Certificate: new SerializableCertificate(certificate))
+                            {
+                                Pkcs7 = Convert.ToBase64String(certificate.Export(X509ContentType.Cert))
+                            };
+                            DatabaseManager.Write(obj, RunId);
+                        }
+                    }
+                    else
+                    {
+                        Log.Debug("Failed to export certificate with OpenSSL.");
+                    }
                 }
+                
             }
             catch (Exception e)
             {
